requirements.txt
streamlit
tensorflow
pillow
matplotlib
numpy
seaborn
scikit-learn